Refactor and modularize aggregation_utils.py
Split BigQueryExecutor, LinkedEdgeGenerator, and ProvenanceSummaryGenerator into separate files under a new 'aggregation' package.

Simplify 'aggregation_utils.py' to act as a thin orchestrator, preserving the public API for backward compatibility with main.py.

Add comprehensive unit tests in 'aggregation_test.py' covering all components.

Format all new/modified code with yapf (Google style) and sort imports with isort.
