+++ pipeline/workflow/ingestion-helper/aggregation_utils.py
@@ -13,647 +13,14 @@
 # limitations under the License.
 
 import logging
-import os
-import time
 from typing import Any, Dict, List, Optional
 
+from aggregation import BigQueryExecutor
+from aggregation import LinkedEdgeGenerator
+from aggregation import ProvenanceSummaryGenerator
 from google.cloud import bigquery
 
 logging.getLogger().setLevel(logging.INFO)
-
-
-def _escape_sql_literal(val: str) -> str:
-    r"""Escapes a string literal for use in nested BigQuery/Spanner queries.
-
-    This is required because the query string travels through two SQL parsers:
-    1. BigQuery parses the EXTERNAL_QUERY double-quoted string literal.
-    2. Spanner parses the resulting inner query's single-quoted string literal.
-
-    To ensure the value is correctly matched and prevent SQL injection:
-    - Backslashes (\) are escaped to 4 backslashes (\\\\) so they survive
-      both decodings (\\\\ -> \\ -> \). Otherwise, they may escape quotes
-      or be interpreted as control characters (like \b becoming backspace).
-    - Double quotes (") are escaped to \\" to prevent terminating BQ string.
-    - Single quotes (') are escaped to '' to prevent terminating Spanner string.
-    """
-    return val.replace('\\', '\\\\\\\\').replace('"', '\\"').replace("'", "''")
-
-
-class BigQueryExecutor:
-    """Handles BigQuery client initialization and query execution."""
-
-    def __init__(self,
-                 connection_id: str,
-                 project_id: str,
-                 instance_id: str,
-                 database_id: str,
-                 location: Optional[str] = None,
-                 run_sequential: bool = True) -> None:
-        self.connection_id = connection_id
-        self.project_id = project_id
-        self.instance_id = instance_id
-        self.database_id = database_id
-        self.location = location
-        self.run_sequential = run_sequential
-        try:
-            self.client = bigquery.Client(project=self.project_id,
-                                          location=self.location)
-        except Exception as e:
-            logging.warning(f"Failed to initialize BigQuery client: {e}")
-            self.client = None
-
-    def get_spanner_destination_uri(self) -> str:
-        """Returns the Spanner destination URI for EXPORT DATA."""
-        return f"https://spanner.googleapis.com/projects/{self.project_id}/instances/{self.instance_id}/databases/{self.database_id}"
-
-    def execute(
-        self,
-        query: str,
-        job_config: Optional[bigquery.QueryJobConfig] = None
-    ) -> bigquery.job.QueryJob:
-        """Submits a query and returns the QueryJob.
-
-        If run_sequential is True, it blocks and waits for the query to complete
-        before returning.
-        """
-        if not self.client:
-            logging.error("BigQuery client not initialized")
-            raise RuntimeError("BigQuery client not initialized")
-
-        logging.info(
-            f"Submitting query (first 100 chars): {query.strip()[:100]}...")
-
-        start_time = time.time()
-        try:
-            query_job = self.client.query(query, job_config=job_config)
-            logging.info(f"Query submitted. Job ID: {query_job.job_id}")
-
-            if self.run_sequential:
-                logging.info(
-                    f"Waiting for Query Job {query_job.job_id} to complete (sequential mode)..."
-                )
-                query_job.result()
-                duration = time.time() - start_time
-                logging.info(
-                    f"Query completed in {duration:.2f}s. Job ID: {query_job.job_id}"
-                )
-
-            return query_job
-        except Exception as e:
-            logging.error(f"Failed to submit or execute query: {e}")
-            raise
-
-    def get_jobs_status(self, job_ids: List[str]) -> Dict[str, Any]:
-        """Returns the overall status of a list of BigQuery jobs."""
-        if not self.client:
-            logging.error("BigQuery client not initialized")
-            raise RuntimeError("BigQuery client not initialized")
-
-        overall_status = "DONE"
-        failed_jobs = []
-        error_message = ""
-
-        for job_id in job_ids:
-            try:
-                job = self.client.get_job(job_id, location=self.location)
-                if job.error_result:
-                    overall_status = "FAILED"
-                    failed_jobs.append(job_id)
-                    error_message += f"Job {job_id} failed: {job.error_result}. "
-                elif job.state != "DONE" and overall_status != "FAILED":
-                    overall_status = "RUNNING"
-            except Exception as e:
-                logging.error(f"Failed to get job status for {job_id}: {e}")
-                overall_status = "FAILED"
-                failed_jobs.append(job_id)
-                error_message += f"Failed to get job {job_id}: {e}. "
-
-        if overall_status == "FAILED":
-            return {
-                "status": overall_status,
-                "error": error_message,
-                "failedJobs": failed_jobs
-            }
-        else:
-            return {"status": overall_status}
-
-
-class LinkedEdgeGenerator:
-    """Generates and ingests linked relationship edges (e.g., transitive closures) into Spanner for faster lookup."""
-
-    def __init__(self,
-                 executor: BigQueryExecutor,
-                 is_base_dc: bool = True) -> None:
-        self.executor = executor
-        self.is_base_dc = is_base_dc
-
-    def run_all(self,
-                import_names: List[str] = None) -> List[bigquery.job.QueryJob]:
-        """Runs all global aggregations asynchronously and returns their jobs."""
-        if not import_names:
-            logging.info("No imports specified. Skipping global aggregations.")
-            return []
-
-        logging.info(f"Running global aggregations for imports: {import_names}")
-
-        jobs = [
-            self.run_linked_contained_in_place(import_names),
-            self.run_linked_member_of(import_names),
-            self.run_linked_member(import_names)
-        ]
-        return jobs
-
-    def run_linked_contained_in_place(self,
-                                      import_names: List[str] = None) -> None:
-        """Expands place containment hierarchies."""
-        if not import_names:
-            return
-
-        dest = self.executor.get_spanner_destination_uri()
-        safe_names = [_escape_sql_literal(name) for name in import_names]
-        prefix = "dc/base/" if self.is_base_dc else ""
-        provenances = [f"'{prefix}{name}'" for name in safe_names]
-        provenance_filter = f" AND provenance IN ({', '.join(provenances)})"
-        gen_graphs_prov = 'dc/base/GeneratedGraphs' if self.is_base_dc else 'GeneratedGraphs'
-
-        query = f"""
-        -- Pull base edges needed for containedInPlace aggregation
-        CREATE OR REPLACE TEMPORARY TABLE `temp_base_contained_in_place` AS
-        SELECT * FROM EXTERNAL_QUERY("{self.executor.connection_id}", 
-          "SELECT subject_id, predicate, object_id FROM Edge WHERE predicate = 'containedInPlace'{provenance_filter}");
-
-        -- Pull existing generated edges to filter them out later
-        CREATE OR REPLACE TEMPORARY TABLE `temp_existing_linked_contained_in_place` AS
-        SELECT * FROM EXTERNAL_QUERY("{self.executor.connection_id}", 
-          "SELECT subject_id, predicate, object_id, provenance FROM Edge WHERE predicate = 'linkedContainedInPlace'");
-
-        CREATE OR REPLACE TEMPORARY TABLE `temp_contained_in_place` AS
-        SELECT subject_id, object_id
-        FROM `temp_base_contained_in_place`;
-
-        EXPORT DATA
-          OPTIONS( uri="{dest}",
-            format='CLOUD_SPANNER',
-            spanner_options = '{{"table": "Edge"}}' ) AS
-        with RECURSIVE Ancestors AS (
-          SELECT
-            subject_id,
-            object_id AS ancestor_place,
-            1 AS level
-          FROM
-            temp_contained_in_place
-          UNION ALL
-
-          SELECT
-            a.subject_id,
-            t.object_id AS ancestor_place,
-            a.level + 1
-          FROM
-            Ancestors AS a
-          JOIN
-            temp_contained_in_place AS t
-            ON a.ancestor_place = t.subject_id
-          WHERE
-            a.level <= 10 -- Limit to 10 levels
-        ),
-        NewEdges AS (
-          SELECT DISTINCT
-            subject_id,
-            'linkedContainedInPlace' as predicate,
-            ancestor_place as object_id,
-            '{gen_graphs_prov}' as provenance
-          FROM
-            Ancestors
-        ),
-        FilteredEdges AS (
-          SELECT
-            subject_id,
-            predicate,
-            object_id,
-            provenance
-          FROM
-            NewEdges n
-          WHERE NOT EXISTS (
-            SELECT 1
-            FROM `temp_existing_linked_contained_in_place` e
-            WHERE n.subject_id = e.subject_id
-              AND n.predicate = e.predicate
-              AND n.object_id = e.object_id
-              AND n.provenance = e.provenance
-          )
-        )
-        SELECT
-          subject_id,
-          predicate,
-          object_id,
-          provenance
-        FROM
-          FilteredEdges
-        """
-        return self.executor.execute(query)
-
-    def run_linked_member_of(self, import_names: List[str] = None) -> None:
-        """Expands membership hierarchies using memberOf and specializationOf."""
-        if not import_names:
-            return
-
-        dest = self.executor.get_spanner_destination_uri()
-        safe_names = [_escape_sql_literal(name) for name in import_names]
-        prefix = "dc/base/" if self.is_base_dc else ""
-        provenances = [f"'{prefix}{name}'" for name in safe_names]
-        provenance_filter = f" AND provenance IN ({', '.join(provenances)})"
-        gen_graphs_prov = 'dc/base/GeneratedGraphs' if self.is_base_dc else 'GeneratedGraphs'
-
-        query = f"""
-        -- Pull base edges needed for memberOf aggregation
-        CREATE OR REPLACE TEMPORARY TABLE `temp_base_member_of` AS
-        SELECT * FROM EXTERNAL_QUERY("{self.executor.connection_id}", 
-          "SELECT subject_id, predicate, object_id FROM Edge WHERE predicate IN ('memberOf', 'specializationOf'){provenance_filter}");
-
-        -- Pull existing generated edges to filter them out later
-        CREATE OR REPLACE TEMPORARY TABLE `temp_existing_linked_member_of` AS
-        SELECT * FROM EXTERNAL_QUERY("{self.executor.connection_id}", 
-          "SELECT subject_id, predicate, object_id, provenance FROM Edge WHERE predicate = 'linkedMemberOf'");
-
-        CREATE OR REPLACE TEMPORARY TABLE `temp_hierarchy` AS
-        SELECT DISTINCT subject_id, predicate, object_id
-        FROM `temp_base_member_of`;
-
-        EXPORT DATA
-          OPTIONS( uri="{dest}",
-            format='CLOUD_SPANNER',
-            spanner_options = '{{"table": "Edge"}}' ) AS
-        WITH RECURSIVE Ancestors AS (
-          SELECT
-            subject_id,
-            object_id AS ancestor,
-            1 AS level
-          FROM
-            temp_hierarchy
-          WHERE
-            predicate = 'memberOf'
-          UNION ALL
-
-          SELECT
-            a.subject_id,
-            t.object_id AS ancestor,
-            a.level + 1
-          FROM
-            Ancestors AS a
-          JOIN
-            temp_hierarchy AS t
-            ON a.ancestor = t.subject_id
-          WHERE
-            a.level <= 20 -- Limit to 20 levels
-            AND t.predicate = 'specializationOf'
-        ),
-        NewEdges AS (
-          SELECT DISTINCT
-            subject_id,
-            'linkedMemberOf' as predicate,
-            ancestor as object_id,
-            '{gen_graphs_prov}' as provenance
-          FROM
-            Ancestors
-        ),
-        FilteredEdges AS (
-          SELECT
-            subject_id,
-            predicate,
-            object_id,
-            provenance
-          FROM
-            NewEdges n
-          WHERE NOT EXISTS (
-            SELECT 1
-            FROM `temp_existing_linked_member_of` e
-            WHERE n.subject_id = e.subject_id
-              AND n.predicate = e.predicate
-              AND n.object_id = e.object_id
-              AND n.provenance = e.provenance
-          )
-        )
-        SELECT
-          subject_id,
-          predicate,
-          object_id,
-          provenance
-        FROM
-          FilteredEdges
-        """
-        return self.executor.execute(query)
-
-    def run_linked_member(self, import_names: List[str] = None) -> None:
-        """Expands topic/SVGP descendants to identify leaf members."""
-        if not import_names:
-            return
-
-        dest = self.executor.get_spanner_destination_uri()
-        safe_names = [_escape_sql_literal(name) for name in import_names]
-        prefix = "dc/base/" if self.is_base_dc else ""
-        provenances = [f"'{prefix}{name}'" for name in safe_names]
-        provenance_filter = f" AND provenance IN ({', '.join(provenances)})"
-        gen_graphs_prov = 'dc/base/GeneratedGraphs' if self.is_base_dc else 'GeneratedGraphs'
-
-        query = f"""
-        -- Pull base edges needed for member aggregation
-        CREATE OR REPLACE TEMPORARY TABLE `temp_base_member` AS
-        SELECT * FROM EXTERNAL_QUERY("{self.executor.connection_id}", 
-          "SELECT subject_id, predicate, object_id FROM Edge WHERE predicate IN ('relevantVariable', 'member'){provenance_filter}");
-
-        -- Pull existing generated edges to filter them out later
-        CREATE OR REPLACE TEMPORARY TABLE `temp_existing_linked_member` AS
-        SELECT * FROM EXTERNAL_QUERY("{self.executor.connection_id}", 
-          "SELECT subject_id, predicate, object_id, provenance FROM Edge WHERE predicate = 'linkedMember'");
-
-        CREATE OR REPLACE TEMPORARY TABLE `temp_topic_hierarchy` AS
-        SELECT DISTINCT subject_id, object_id
-        FROM `temp_base_member`
-        WHERE (subject_id LIKE 'dc/topic%' OR subject_id LIKE 'dc/svpg%');
-
-        EXPORT DATA
-          OPTIONS( uri="{dest}",
-            format='CLOUD_SPANNER',
-            spanner_options = '{{"table": "Edge"}}' ) AS
-        WITH RECURSIVE Descendants AS (
-          SELECT
-            subject_id,
-            object_id AS descendant,
-            1 AS level
-          FROM
-            temp_topic_hierarchy
-          UNION ALL
-
-          SELECT
-            d.subject_id,
-            t.object_id AS descendant,
-            d.level + 1
-          FROM
-            Descendants AS d
-          JOIN
-            temp_topic_hierarchy AS t
-            ON d.descendant = t.subject_id
-          WHERE
-            d.level <= 20 -- Limit to 20 levels
-        ),
-        NewEdges AS (
-          SELECT DISTINCT
-            descendant as subject_id,
-            'linkedMember' as predicate,
-            subject_id as object_id,
-            '{gen_graphs_prov}' as provenance
-          FROM
-            Descendants
-          WHERE subject_id LIKE 'dc/topic%'
-          AND descendant NOT LIKE 'dc/topic%'
-          AND descendant NOT LIKE 'dc/svpg%'
-        ),
-        FilteredEdges AS (
-          SELECT
-            subject_id,
-            predicate,
-            object_id,
-            provenance
-          FROM
-            NewEdges n
-          WHERE NOT EXISTS (
-            SELECT 1
-            FROM `temp_existing_linked_member` e
-            WHERE n.subject_id = e.subject_id
-              AND n.predicate = e.predicate
-              AND n.object_id = e.object_id
-              AND n.provenance = e.provenance
-          )
-        )
-        SELECT
-          subject_id,
-          predicate,
-          object_id,
-          provenance
-        FROM
-          FilteredEdges
-        """
-        return self.executor.execute(query)
-
-
-class ProvenanceSummaryGenerator:
-    """Contains the SQL queries to generate ProvenanceSummary in the Cache table."""
-
-    def __init__(self,
-                 executor: BigQueryExecutor,
-                 is_base_dc: bool = True) -> None:
-        self.executor = executor
-        self.is_base_dc = is_base_dc
-
-    def run_all(self, import_names: List[str]) -> List[bigquery.job.QueryJob]:
-        """Runs all provenance summary generation asynchronously and returns their jobs."""
-        if not import_names:
-            logging.info("No imports specified. Skipping cache aggregations.")
-            return []
-
-        logging.info(
-            f"Running provenance summary generation for imports: {import_names}"
-        )
-        return [self.run_provenance_summary_aggregation(import_names)]
-
-    def run_provenance_summary_aggregation(self,
-                                           import_names: List[str]) -> None:
-        """Calculates ProvenanceSummary for all variables and populates the Cache table."""
-        if not import_names:
-            return
-
-        dest = self.executor.get_spanner_destination_uri()
-        connection_id = self.executor.connection_id
-
-        safe_names = [_escape_sql_literal(name) for name in import_names]
-        # Format import names for the SQL IN clause
-        imports_str = ", ".join([f"'{name}'" for name in safe_names])
-        provenance_dcid_expr = "CONCAT('dc/base/', raw.import_name)" if self.is_base_dc else "raw.import_name"
-
-        query = f"""
-        -- Step 1: Fetch Observation rows for the specific import
-        -- We cast 'observations' to STRING to avoid the PROTO error.
-        CREATE OR REPLACE TEMPORARY TABLE `temp_obs_raw` AS
-        SELECT 
-          variable_measured, 
-          observation_about, 
-          facet_id, 
-          import_name,
-          observation_period,
-          measurement_method,
-          unit,
-          scaling_factor,
-          is_dc_aggregate,
-          observations_json
-        FROM EXTERNAL_QUERY("{connection_id}", 
-          '''SELECT 
-               variable_measured, 
-               observation_about, 
-               facet_id, 
-               import_name,
-               observation_period,
-               measurement_method,
-               unit,
-               scaling_factor,
-               is_dc_aggregate,
-               IF(ARRAY_LENGTH(observations.values) > 0,
-                 (
-                   SELECT CONCAT('{{"values":[', STRING_AGG(FORMAT('{{"key":"%s","value":"%s"}}', entry.key, entry.value), ','), ']}}')
-                   FROM UNNEST(observations.values) as entry
-                 ),
-                 NULL
-               ) as observations_json
-             FROM Observation
-             WHERE import_name IN ({imports_str}) ''');
-
-        -- Step 2: Fetch ALL Node names (Narrow selection to reduce data transfer)
-        CREATE OR REPLACE TEMPORARY TABLE `temp_node_names` AS
-        SELECT subject_id, name 
-        FROM EXTERNAL_QUERY("{connection_id}",
-          "SELECT subject_id, name FROM Node WHERE name IS NOT NULL");
-
-        -- Step 3: Fetch ALL typeOf edges (Narrow selection)
-        CREATE OR REPLACE TEMPORARY TABLE `temp_type_edges` AS
-        SELECT subject_id, object_id as place_type
-        FROM EXTERNAL_QUERY("{connection_id}", 
-          "SELECT subject_id, object_id FROM Edge WHERE predicate = 'typeOf'");
-
-        -- Step 4: Join and Flatten in BigQuery
-        CREATE OR REPLACE TEMPORARY TABLE `temp_prepared` AS
-        SELECT 
-          raw.variable_measured,
-          raw.observation_about,
-          raw.facet_id,
-          raw.import_name,
-          raw.observation_period,
-          raw.measurement_method,
-          raw.unit,
-          raw.scaling_factor,
-          raw.is_dc_aggregate,
-          JSON_VALUE(v, '$.key') as date_val,
-          SAFE_CAST(JSON_VALUE(v, '$.value') AS FLOAT64) as value_num,
-          {provenance_dcid_expr} as provenance_dcid,
-          nodes.name as place_name,
-          edges.place_type
-        FROM `temp_obs_raw` raw
-        CROSS JOIN UNNEST(JSON_EXTRACT_ARRAY(SAFE.PARSE_JSON(observations_json), '$.values')) as v
-        LEFT JOIN `temp_node_names` nodes ON raw.observation_about = nodes.subject_id
-        LEFT JOIN `temp_type_edges` edges ON raw.observation_about = edges.subject_id;
-
-        -- Step 5: Aggregate Place Type Summaries
-        CREATE OR REPLACE TEMPORARY TABLE `temp_place_type_summary` AS
-        SELECT
-          variable_measured,
-          provenance_dcid,
-          facet_id,
-          place_type,
-          COUNT(DISTINCT observation_about) as place_count,
-          MIN(value_num) as min_val,
-          MAX(value_num) as max_val,
-          ARRAY_AGG(
-            STRUCT(observation_about as dcid, place_name as name)
-            ORDER BY observation_about LIMIT 3
-          ) as top_places
-        FROM `temp_prepared`
-        WHERE place_type IS NOT NULL
-        GROUP BY variable_measured, provenance_dcid, facet_id, place_type;
-
-        -- Step 6: Final aggregation and export to Cache
-        EXPORT DATA
-          OPTIONS( uri="{dest}",
-            format='CLOUD_SPANNER',
-            spanner_options = '{{"table": "Cache"}}' ) AS
-        WITH facet_base AS (
-          SELECT 
-            variable_measured, provenance_dcid, facet_id,
-            ANY_VALUE(import_name) as import_name,
-            ANY_VALUE(measurement_method) as measurement_method,
-            ANY_VALUE(observation_period) as observation_period,
-            ANY_VALUE(unit) as unit,
-            ANY_VALUE(scaling_factor) as scaling_factor,
-            ANY_VALUE(is_dc_aggregate) as is_dc_aggregate,
-            MIN(date_val) as min_date,
-            MAX(date_val) as max_date,
-            MIN(value_num) as facet_min,
-            MAX(value_num) as facet_max,
-            COUNT(*) as facet_obs_count,
-            COUNT(DISTINCT observation_about) as facet_ts_count
-          FROM `temp_prepared`
-          GROUP BY variable_measured, provenance_dcid, facet_id
-        ),
-        facet_summaries AS (
-          SELECT 
-            fb.variable_measured,
-            fb.provenance_dcid,
-            fb.facet_id,
-            fb.import_name,
-            fb.measurement_method,
-            fb.observation_period,
-            fb.unit,
-            fb.scaling_factor,
-            fb.is_dc_aggregate,
-            fb.min_date,
-            fb.max_date,
-            fb.facet_min,
-            fb.facet_max,
-            fb.facet_obs_count,
-            fb.facet_ts_count,
-            ARRAY_AGG(STRUCT(pts.place_type, pts.place_count, pts.min_val, pts.max_val, pts.top_places)) as pt_summaries
-          FROM facet_base fb
-          LEFT JOIN `temp_place_type_summary` pts USING (variable_measured, provenance_dcid, facet_id)
-          GROUP BY 
-            variable_measured, provenance_dcid, facet_id, import_name, measurement_method,
-            observation_period, unit, scaling_factor, is_dc_aggregate, min_date, max_date,
-            facet_min, facet_max, facet_obs_count, facet_ts_count
-        )
-        SELECT
-          'ProvenanceSummary' as type,
-          variable_measured as key,
-          provenance_dcid as provenance,
-          JSON_OBJECT(
-            'import_name', ANY_VALUE(import_name),
-            'observation_count', CAST(SUM(facet_obs_count) AS FLOAT64),
-            'time_series_count', CAST(SUM(facet_ts_count) AS FLOAT64),
-            'series_summary', ARRAY_AGG(
-              JSON_OBJECT(
-                'series_key', JSON_OBJECT(
-                  'measurement_method', measurement_method,
-                  'observation_period', observation_period,
-                  'unit', unit,
-                  'scaling_factor', scaling_factor,
-                  'is_dc_aggregate', COALESCE(is_dc_aggregate, false)
-                ),
-                'earliest_date', min_date,
-                'latest_date', max_date,
-                'min_value', facet_min,
-                'max_value', facet_max,
-                'observation_count', CAST(facet_obs_count AS FLOAT64),
-                'time_series_count', CAST(facet_ts_count AS FLOAT64),
-                'place_type_summary', (
-                  SELECT IF(ARRAY_LENGTH(keys) > 0, JSON_OBJECT(keys, vals), NULL)
-                  FROM (
-                    SELECT 
-                      ARRAY_AGG(place_type) as keys,
-                      ARRAY_AGG(JSON_OBJECT(
-                        'place_count', place_count,
-                        'min_value', min_val,
-                        'max_value', max_val,
-                        'top_places', (
-                          SELECT ARRAY_AGG(JSON_OBJECT('dcid', tp.dcid, 'name', tp.name))
-                          FROM UNNEST(top_places) tp
-                        )
-                      )) as vals
-                    FROM UNNEST(pt_summaries)
-                    WHERE place_type IS NOT NULL
-                  )
-                )
-              )
-            )
-          ) as value
-        FROM facet_summaries
-        GROUP BY variable_measured, provenance_dcid;
-        """
-        return self.executor.execute(query)
 
 
 class AggregationUtils:
